Moving script utils
