+++ examples/simple_static_task/static_test_script.py
@@ -8,7 +8,7 @@ import os
 from mephisto.core.operator import Operator
 from mephisto.core.utils import get_root_dir
 from mephisto.server.blueprints.static_task.static_html_blueprint import BLUEPRINT_TYPE
-from mephisto.utils.scripts import load_db_and_process_config
+from mephisto.tools.scripts import load_db_and_process_config
 
 import hydra
 from omegaconf import DictConfig
