+++ examples/static_react_task/run_task.py
@@ -9,7 +9,7 @@ import shutil
 import subprocess
 from mephisto.core.operator import Operator
 from mephisto.core.utils import get_root_dir
-from mephisto.utils.scripts import load_db_and_process_config
+from mephisto.tools.scripts import load_db_and_process_config
 from mephisto.server.blueprints.static_react_task.static_react_blueprint import (
     BLUEPRINT_TYPE,
 )
